+++ packages/boilerplate/theme/nuxt.config.js
@@ -19,7 +19,14 @@ export default {
     link: [
       { rel: 'icon',
         type: 'image/x-icon',
-        href: '/favicon.ico' }
+        href: '/favicon.ico'
+      }, {
+        rel: 'preconnect',
+        href: 'https://fonts.googleapis.com'
+      }, {
+        rel: 'preconnect',
+        href: 'https://fonts.gstatic.com'
+      }
     ]
   },
   loading: { color: '#fff' },
@@ -29,6 +36,7 @@ export default {
   buildModules: [
     // to core
     '@nuxt/typescript-build',
+    '@nuxtjs/style-resources',
     ['@vue-storefront/nuxt', {
       // @core-development-only-start
       coreDevelopment: true,
@@ -80,6 +88,9 @@ export default {
       }
     }
   },
+  styleResources: {
+    scss: [require.resolve('@storefront-ui/shared/styles/_helpers.scss', { paths: [process.cwd()] })]
+  },
   build: {
     transpile: [
       'vee-validate/dist/rules'
@@ -93,5 +104,14 @@ export default {
         })
       })
     ]
+  },
+  router: {
+    scrollBehavior (_to, _from, savedPosition) {
+      if (savedPosition) {
+        return savedPosition;
+      } else {
+        return { x: 0, y: 0 };
+      }
+    }
   }
 };