+++ packages/checkout-com/nuxt/index.js
@@ -3,7 +3,6 @@ import proxyMiddleware from '@vue-storefront/checkout-com/nuxt/proxyMiddleware';
 
 const defaultPaymentMethods = {
   cc: true,
-  paypal: true,
   klarna: false
 };