+++ packages/commercetools/theme/components/Checkout/UserShippingAddresses.vue
@@ -66,8 +66,6 @@ export default {
 </script>
 
 <style lang="scss">
-@import "~@storefront-ui/vue/styles";
-
 .shipping-addresses {
   @include for-desktop {
     display: grid;