+++ packages/core/nuxt-theme-module/theme/components/RelatedProducts.vue
@@ -49,9 +49,8 @@ export default {
   }
 };
 </script>
-<style lang="scss" scoped>
-@import "~@storefront-ui/vue/styles";
 
+<style lang="scss" scoped>
 .section {
   margin-top: var(--spacer-base);
 }