+++ packages/core/nuxt-theme-module/theme/pages/MyAccount/MyNewsletter.vue
@@ -37,6 +37,7 @@
     </SfTab>
   </SfTabs>
 </template>
+
 <script>
 import { SfTabs, SfCheckbox, SfButton, SfLink } from '@storefront-ui/vue';
 export default {
@@ -52,9 +53,8 @@ export default {
   }
 };
 </script>
-<style lang='scss' scoped>
-@import "~@storefront-ui/vue/styles";
 
+<style lang='scss' scoped>
 .tab-orphan {
   @include for-mobile {
     --tabs-title-display: none;