+++ packages/shopify/theme/components/Filters.vue
@@ -1,3 +1,0 @@
-<template>
- <div>This component should be overriden in integration-specific theme.</div>
-</template>