+++ packages/core/docs/checkout-com/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.0.9 (not released)
+- Sofort support ([#5158](https://github.com/DivanteLtd/vue-storefront/issues/5158))
+
+## 0.0.8
+
+- Fixed proxied endpoints URLs ([#5117](https://github.com/DivanteLtd/vue-storefront/pull/5117))
+
 ## 0.0.7
 
 - Removed `ckoWebHookUrl` ([#4910](https://github.com/DivanteLtd/vue-storefront/issues/4910))